+++ src/__tests__/Header.test.jsx
@@ -1,20 +1,18 @@
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { beforeEach, describe, expect, it } from "vitest";
+import { MemoryRouter } from "react-router-dom";
+import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
 import Header from "../components/Header";
-import routes from "../routes/routes";
 /* Testing components that renders a <Link> or a <Route>
  * https://v5.reactrouter.com/web/guides/testing
  */
-const router = createBrowserRouter(routes);
 
 describe("Header component", () => {
   // Cannot test header with Link components from react-router-dom
   it("The Header component is rendered", () => {
     render(
-      <RouterProvider router={router}>
+      <MemoryRouter>
         <Header />
-      </RouterProvider>
+      </MemoryRouter>
     );
     const header = screen.getByRole("banner");
     expect(header).toBeInTheDocument();
@@ -22,9 +20,9 @@ describe("Header component", () => {
 
   it("A nav element is rendered", () => {
     render(
-      <RouterProvider router={router}>
+      <MemoryRouter>
         <Header />
-      </RouterProvider>
+      </MemoryRouter>
     );
     const nav = screen.getByRole("navigation");
     expect(nav).toBeInTheDocument();
@@ -32,9 +30,9 @@ describe("Header component", () => {
 
   it("There are exactly 2 list elements inside the nav element", () => {
     render(
-      <RouterProvider router={router}>
+      <MemoryRouter>
         <Header />
-      </RouterProvider>
+      </MemoryRouter>
     );
     const lists = screen.getAllByRole("list");
     expect(lists.length).equal(2);