+++ src/components/Product.jsx
@@ -1,5 +1,27 @@
+import { useLocation, useOutletContext } from "react-router-dom";
+
 const Product = () => {
-  return <section>Product</section>;
+  const { state } = useLocation();
+  const { product } = state;
+  console.log(state);
+  return (
+    <section>
+      <article>
+        <h4>{product.title}</h4>
+        <picture>
+          <img
+            style={{ width: "200px" }}
+            src={product.image}
+            loading="lazy"
+            alt="#"
+          />
+        </picture>
+        <div>
+          <p>{product.description}</p>
+        </div>
+      </article>
+    </section>
+  );
 };
 
 export default Product;